Feature/meta

See merge request deepexi/xpaas-backend/commodity-domain/deepexi-channel-center/deepexi-channel-center!34

+++ xpaas-channel-center-channel/src/main/java/com/deepexi/channel/domain/distributor/DistributorDO.java
@@ -37,7 +37,7 @@ public class DistributorDO extends SuperEntity {
      * 上级经销商ID
      */
     @ApiModelProperty("上级经销商ID")
-    private int parent_id;
+    private Long parent_id;
 
     /**
      * 经销商类型 1 厂商 2 经销商
