+++ xpaas-channel-center-channel/src/main/java/com/deepexi/channel/domain/distributor/DistributorVO.java
@@ -62,7 +62,7 @@ public class DistributorVO extends AbstractObject {
      * 上级经销商ID
      */
     @ApiModelProperty("上级经销商ID")
-    private int parent_id;
+    private Long parent_id;
 
     /**
      * 经销商类型 1 厂商 2 经销商
